move variants into the kits folder
